Force the OSX compatibility level
We are building mac binaries on modern M1 macs. In order to make sure
that the binaries produced are still usable on older systems, we need
to enforce desired compatibility level. Now we are producing binaries
which should be compatible with at least OSX 10.13 (x86_64) and
OSX 11.0 (arm64).

CCISSUE: SQUISH-14692